Added About page (work in progress)
Work in progress, but I really need to sleep. Team section of About
should be expanded upon; more info would be nice. Would like to read a
unified toolList JSON object, then, using this information, also pull
in license texts to be displayed in the Open Source Licenses section
of About; this likely requires all tools to be moved to their own
directories and have LICENSE files included with each.