fixed issue with feedback containing encoded chars re: #4
This was done by adding another property to the feedback, containing the
he.decode'd targets: feedbackItem.decodedTarget. This is now displayed
wherever there were improperly decoded targets (particularly in
popovers).

Programming feedback was also being improperly highlighted. This
resulted in an unfortunate error where a text button would be injected
in the middle of the encoding for a horizontal tab char. This has been
solved by selecting the entire line as the target for highlighting,
rather than a character selection.