laid groundwork for new student-avatar handling system
Removed all student-avatarFileName keys from databases; this field was
made unnecessary by commit 0d26f93. Student avatars are now stored under
ifs/shared/img/users/$UID/avatar.png

This commit revises the preferences form to allow for a user to pick an
image file and upload it to the server. As of yet, the server does not
handle this file.

This commit also lays out a front-end foundation that allows for future
tools to (with some additional hacking would be required) take files as
parameters for running. This might be useful if a tool wishes to take
for example, an *.ini formatted config file.