add course enrolment/unenrolment in preferences
This commit adds the studentCoursesDB module, which contains functions
for accessing, inserting, and deleting rows in the student_class table.
Some better file organization might be required, but for now the
studentCoursesDB module is located under app/components/StudentProfile
as it pertains directly to student-users. The module is made use of by
the Preferences component.

Other changes include minor code/style review; and an update in
Databases/dbHelpers.js that corrected the update() function to call
buildUpdate instead buildDelete as it was previously.