added post-registration set up form, moved courses out of prefs
Important changes to note here, I tried very hard to do this without
Angular, but was pretty well unsuccessful. The methods in the new Setup
module are very similar to those of Preferences.

This commit added an 'optedIn' column to the users database. This is a
boolean that is toggled by the form at /setup.