users are not allowed to log in without email verification
I have added an additional database check in passport's local-login strategy,
which looks for an entry in the user_registration table. If a row has been
returned that means that the user has completed verification via the emailed
link generated by the Verify module. If no row is returned, the user has not
completed registration, and login is disallowed.

Also in this commit: the /setup route now sets a database value that indicates
whether or not a user has completed setup; if the setup has not been
completed, the user is redirected to this route upon login until it is
completed (they may still manually navigate to the rest of the site though, as
@ian-james has suggested --- we must not block users from using the software
entirely). Once setup has been completed once, they will never be redirected
to /setup again, though once again they may navigate there manually to change
course enrolment, etc.