finished password reset and account verifications re: #69
in this commit: password reset has been enabled and /forgot /reset and /verify
routes have been restricted to users which are not logged in; several checks
are in place to ensure that a user's password should be maliciously reset,
including email link verification and redundant database checks before taking
action to change a user account. /register and /reset pages look more
consistent now as well.